requirements.txt
coverage==5.5
eventlet==0.31.0
Flask==2.2.5
Flask-Babel==2.0.0
flask-caching==1.10.1
Flask-Cors==3.0.10
Flask-Migrate==3.0.1
Flask-RESTful==0.3.9
flask-redis==0.4.0
Flask-Script==2.0.6
Flask-SocketIO==4.3.1
python-engineio==3.13.2
python-socketio==4.6.0
Flask-SQLAlchemy==2.5.1
flask-swagger-ui==3.36.0
marshmallow==3.12.1
mockredis==0.1.3.dev0
pyaml==20.4.0
pytest==8.0.0
PyMySQL==1.0.2
psycopg2-binary==2.8.6
requests==2.25.1
rq==1.8.1
SQLAlchemy==1.4.18
SQLAlchemy-i18n==1.1.0
werkzeug==2.3.7
croniter==2.0.1
mock==5.1.0
pytest-asyncio==0.23.5
aiohttp==3.9.5
python-dotenv==1.0.1
pytest-dotenv==0.5.2
aioresponses==0.7.6